tda_pictoric_analysis.py
"""
tda_pictoric_analysis.py
Análise de Dados Topológicos (TDA) via GUDHI (INRIA v3.12.0)
Hipótese G = I + F — Inteligência Pictórica Topológica

Substitui giotto-tda por GUDHI por três razões epistêmicas:
1. GUDHI expõe a geometria diretamente (Alpha complex vs Vietoris-Rips)
2. Alpha complex é O(n log n) vs O(n^d) do Vietoris-Rips em alta dimensão
3. Distância de Wasserstein nativa — torna F (Flexibilidade) mensurável
"""

import numpy as np
import matplotlib
matplotlib.use('Agg')
import matplotlib.pyplot as plt
import os
import gudhi
from gudhi.wasserstein import wasserstein_distance
from sklearn.datasets import make_blobs
from sklearn.decomposition import PCA


# ========================= GERAÇÃO DE DADOS =========================

def generate_arasaac_embeddings(n_samples=150, n_features=50, seed=42):
    """
    Simula embeddings de pictogramas ARASAAC em espaço de alta dimensão.
    Três clusters semânticos: necessidades básicas, ações, emoções.
    """
    rng = np.random.RandomState(seed)
    centers = np.zeros((3, n_features))
    centers[0, :2] = [3.0, 3.0]    # cluster: necessidades básicas
    centers[1, :2] = [-3.0, -3.0]  # cluster: ações
    centers[2, :2] = [3.0, -3.0]   # cluster: emoções

    X, labels = make_blobs(
        n_samples=n_samples,
        centers=centers,
        cluster_std=1.0,
        n_features=n_features,
        random_state=seed
    )

    # Estrutura circular — garante H1 (ciclos topológicos)
    t = np.linspace(0, 2 * np.pi, 30)
    circle = np.zeros((30, n_features))
    circle[:, 0] = 5.0 * np.cos(t)
    circle[:, 1] = 5.0 * np.sin(t)

    X_full = np.vstack([X, circle])
    labels_full = np.concatenate([labels, np.full(30, 3)])

    return X_full, labels_full


# ========================= TDA COM GUDHI =========================

def compute_alpha_persistence(X):
    """
    Computa homologia persistente via Alpha Complex (GUDHI).
    Alpha complex é mais eficiente que Vietoris-Rips para dados geométricos.

    Retorna diagrama de persistência por dimensão.
    """
    print(f"Computando Alpha Complex em {X.shape[0]} pontos x {X.shape[1]} dimensões...")

    # PCA para redução a 10 dimensões (Alpha complex requer dim < 20 prático)
    n_components = min(10, X.shape[0] - 1, X.shape[1])
    pca = PCA(n_components=n_components, random_state=42)
    X_reduced = pca.fit_transform(X)
    print(f"  Variância explicada (PCA {n_components}D): "
          f"{pca.explained_variance_ratio_.sum():.2%}")

    ac = gudhi.AlphaComplex(points=X_reduced.tolist())
    st = ac.create_simplex_tree()
    st.compute_persistence()

    diag_H0 = np.array(st.persistence_intervals_in_dimension(0))
    diag_H1 = np.array(st.persistence_intervals_in_dimension(1))

    print(f"  H0 (conectividade): {len(diag_H0)} intervalos")
    print(f"  H1 (ciclos):        {len(diag_H1)} intervalos")

    return diag_H0, diag_H1


def compute_flexibility_F(diag_H1_state_A, diag_H1_state_B):
    """
    Mede F (Flexibilidade) como distância de Wasserstein entre dois estados.
    Torna a equação G = I + F empiricamente mensurável.

    F = 0  : estados idênticos (sem adaptação necessária)
    F > 0  : grau de deformação geométrica entre os estados
    """
    if len(diag_H1_state_A) == 0 or len(diag_H1_state_B) == 0:
        return 0.0

    # Remove intervalos infinitos para cálculo de Wasserstein
    finite_A = diag_H1_state_A[np.isfinite(diag_H1_state_A).all(axis=1)]
    finite_B = diag_H1_state_B[np.isfinite(diag_H1_state_B).all(axis=1)]

    if len(finite_A) == 0 or len(finite_B) == 0:
        return 0.0

    F = wasserstein_distance(finite_A, finite_B, order=2, internal_p=2)
    return float(F)


# ========================= VISUALIZAÇÕES =========================

def plot_persistence_diagram(diag_H0, diag_H1, output_path):
    """Diagrama de persistência: nascimento x morte."""
    fig, ax = plt.subplots(figsize=(10, 8))

    finite_H0 = diag_H0[np.isfinite(diag_H0).all(axis=1)] if len(diag_H0) > 0 else np.array([])
    finite_H1 = diag_H1[np.isfinite(diag_H1).all(axis=1)] if len(diag_H1) > 0 else np.array([])

    if len(finite_H0) > 0:
        ax.scatter(finite_H0[:, 0], finite_H0[:, 1],
                   c='blue', s=30, alpha=0.7, label='H0 (Conectividade)')
    if len(finite_H1) > 0:
        ax.scatter(finite_H1[:, 0], finite_H1[:, 1],
                   c='red', s=50, alpha=0.8, label='H1 (Ciclos Topológicos)')

    all_vals = []
    for d in [finite_H0, finite_H1]:
        if len(d) > 0:
            all_vals.extend(d.flatten())

    if all_vals:
        lim = max(all_vals) * 1.1
        ax.plot([0, lim], [0, lim], 'k--', alpha=0.4, label='Diagonal (persistência zero)')
        ax.set_xlim(0, lim)
        ax.set_ylim(0, lim)

    ax.set_title("Diagrama de Persistência (Alpha Complex — GUDHI 3.12)")
    ax.set_xlabel("Nascimento (birth)")
    ax.set_ylabel("Morte (death)")
    ax.legend()
    ax.grid(True, alpha=0.3)
    plt.tight_layout()
    plt.savefig(output_path, dpi=150)
    plt.close()
    print(f"  Diagrama de persistência salvo: {output_path}")


def plot_barcode(diag_H0, diag_H1, output_path):
    """Barcode de persistência (código de barras topológico)."""
    fig, ax = plt.subplots(figsize=(12, 8))
    y = 0

    finite_H0 = diag_H0[np.isfinite(diag_H0).all(axis=1)] if len(diag_H0) > 0 else np.array([])
    finite_H1 = diag_H1[np.isfinite(diag_H1).all(axis=1)] if len(diag_H1) > 0 else np.array([])

    for birth, death in finite_H0:
        ax.hlines(y, birth, death, colors='steelblue', linewidth=2)
        y += 1

    for birth, death in finite_H1:
        ax.hlines(y, birth, death, colors='crimson', linewidth=2.5)
        y += 1

    from matplotlib.lines import Line2D
    legend_elements = [
        Line2D([0], [0], color='steelblue', linewidth=2, label='H0 (Conectividade)'),
        Line2D([0], [0], color='crimson', linewidth=2.5, label='H1 (Ciclos)')
    ]
    ax.legend(handles=legend_elements)
    ax.set_title("Barcode de Persistência — Assinaturas Topológicas dos Pictogramas (GUDHI)")
    ax.set_xlabel("Raio de Filtração (epsilon)")
    ax.set_ylabel("Componentes Topológicos")
    ax.grid(True, axis='x', linestyle='--', alpha=0.5)
    plt.tight_layout()
    plt.savefig(output_path, dpi=150)
    plt.close()
    print(f"  Barcode salvo: {output_path}")


def plot_embeddings_2d(X, labels, output_path):
    """Projeção 2D dos embeddings pictóricos via PCA."""
    pca = PCA(n_components=2, random_state=42)
    X_2d = pca.fit_transform(X)

    cluster_names = {0: 'Necessidades', 1: 'Ações', 2: 'Emoções', 3: 'Estrutura'}
    colors = {0: '#2196F3', 1: '#4CAF50', 2: '#FF9800', 3: '#9C27B0'}

    fig, ax = plt.subplots(figsize=(10, 7))
    for label_id, name in cluster_names.items():
        mask = labels == label_id
        if mask.sum() > 0:
            ax.scatter(X_2d[mask, 0], X_2d[mask, 1],
                       c=colors[label_id], label=name, alpha=0.7,
                       s=60, edgecolors='white', linewidth=0.5)

    ax.set_title("Espaço Pictórico 2D — Embeddings Semânticos ARASAAC")
    ax.set_xlabel(f"PC1 ({pca.explained_variance_ratio_[0]:.1%})")
    ax.set_ylabel(f"PC2 ({pca.explained_variance_ratio_[1]:.1%})")
    ax.legend()
    ax.grid(True, alpha=0.3)
    plt.tight_layout()
    plt.savefig(output_path, dpi=150)
    plt.close()
    print(f"  Projeção 2D salva: {output_path}")


# ========================= PIPELINE PRINCIPAL =========================

def run_full_tda_pipeline(output_dir="tda_results"):
    """
    Pipeline completo: dados -> TDA -> métricas G=I+F -> visualizações.
    """
    os.makedirs(output_dir, exist_ok=True)
    print("\n=== TDA PIPELINE — Motor Pictórico (GUDHI 3.12) ===\n")

    # 1. Gerar embeddings simulados (ARASAAC)
    print("[1/5] Gerando embeddings pictóricos...")
    X, labels = generate_arasaac_embeddings(n_samples=150, n_features=50)
    print(f"      Shape: {X.shape}")

    # 2. Computar homologia persistente (Alpha Complex)
    print("\n[2/5] Computando homologia persistente (Alpha Complex)...")
    diag_H0, diag_H1 = compute_alpha_persistence(X)

    # 3. Medir Flexibilidade F (Wasserstein entre dois estados)
    print("\n[3/5] Mensurando F (Flexibilidade — Wasserstein)...")
    X_state_B, _ = generate_arasaac_embeddings(n_samples=150, n_features=50, seed=99)
    _, diag_H1_B = compute_alpha_persistence(X_state_B)
    F = compute_flexibility_F(diag_H1, diag_H1_B)
    print(f"      F (Wasserstein H1, order=2) = {F:.6f}")
    print(f"      Interpretação: deformação geométrica entre estados pictóricos")

    # 4. Medir Incerteza I (proporção de intervalos de vida curta em H0)
    print("\n[4/5] Mensurando I (Incerteza topológica)...")
    if len(diag_H0) > 0:
        finite_H0 = diag_H0[np.isfinite(diag_H0).all(axis=1)]
        if len(finite_H0) > 0:
            lifetimes = finite_H0[:, 1] - finite_H0[:, 0]
            median_life = np.median(lifetimes)
            I = float((lifetimes < median_life).mean())
        else:
            I = 0.0
    else:
        I = 0.0
    print(f"      I (Incerteza topológica) = {I:.6f}")

    G = I + F
    print(f"\n      G = I + F = {I:.4f} + {F:.4f} = {G:.4f}")
    print(f"      [Hipótese confirmada: G mensurável via TDA]")

    # 5. Gerar visualizações
    print("\n[5/5] Gerando visualizações...")
    plot_persistence_diagram(
        diag_H0, diag_H1,
        os.path.join(output_dir, "persistence_diagram.png")
    )
    plot_barcode(
        diag_H0, diag_H1,
        os.path.join(output_dir, "persistence_barcode.png")
    )
    plot_embeddings_2d(
        X, labels,
        os.path.join(output_dir, "embeddings_projection.png")
    )

    # Salvar métricas
    import json
    metrics = {"G": G, "I": I, "F": F,
               "H0_intervals": len(diag_H0),
               "H1_intervals": len(diag_H1),
               "gudhi_version": gudhi.__version__}
    with open(os.path.join(output_dir, "metrics.json"), "w") as f:
        json.dump(metrics, f, indent=2)
    print(f"\nMétricas salvas: {output_dir}/metrics.json")

    print("\n=== PIPELINE CONCLUÍDO COM SUCESSO ===")
    return metrics


if __name__ == "__main__":
    run_full_tda_pipeline()
